+++ src/index.js
@@ -1,12 +1,19 @@
+//-- UTILS -------------------------------------------------------------------------------------------------------------
 import _isReactComponent from './util/is-react-component';
 export { _isReactComponent as isReactComponent };
 
+//-- WRAPPERS ----------------------------------------------------------------------------------------------------------
 import _RenderCounter from './wrappers/render-counter';
 export { _RenderCounter as RenderCounter };
 
+//-- COMPONENTS --------------------------------------------------------------------------------------------------------
 import _Lorem from './components/react-lorem';
 export { _Lorem as Lorem };
+
+//-- COMPONENTS.LAYOUT -------------------------------------------------------------------------------------------------
 import _VCenter from './components/layout/react-vcenter';
 export { _VCenter as VCenter };
 import _AboveFold from './components/layout/react-above-fold';
 export { _AboveFold as AboveFold };
+import _RatioPlaceholder from './components/layout/react-ratio-placeholder';
+export { _RatioPlaceholder as RatioPlaceholder };